refactor(go): simplify execute, predicate, and lower internals
NFC. Consolidates assertion/capture orchestration into clearer runner\npaths, splits lower translation helpers by concern, and removes dead\nwrapper code.

Also restores reusable predicate evaluation state in execution to\navoid per-call evaluator rebuilds and duplicate regex-capture logic.
