+++ internal/rq/assert/predicate.go
@@ -5,6 +5,25 @@ import (
 	"github.com/jacoelho/rq/internal/rq/predicate"
 )
 
+type Evaluator struct {
+	predicate *predicate.Evaluator
+}
+
+func NewEvaluator() *Evaluator {
+	return &Evaluator{
+		predicate: predicate.NewEvaluator(),
+	}
+}
+
+func (e *Evaluator) Evaluate(actual any, input model.Predicate) (bool, error) {
+	expr, err := BuildExpr(input)
+	if err != nil {
+		return false, err
+	}
+
+	return e.predicate.Evaluate(expr, actual)
+}
+
 // BuildExpr converts parser predicate input into a validated predicate expression.
 func BuildExpr(input model.Predicate) (predicate.Expr, error) {
 	op, err := predicate.ParseOperator(input.Operation)
@@ -31,10 +50,5 @@ func Validate(input model.Predicate) error {
 }
 
 func Evaluate(actual any, input model.Predicate) (bool, error) {
-	expr, err := BuildExpr(input)
-	if err != nil {
-		return false, err
-	}
-
-	return predicate.EvaluateExpr(expr, actual)
+	return NewEvaluator().Evaluate(actual, input)
 }
